funcion imprimir orden y alteracion del canpo cantidad
se altera el campo cantidad en transaccionOrden para poder tener cantidades decimales.

+++ lugrascolv2/facturacion/models.py
@@ -126,7 +126,7 @@ class TransaccionOrden(models.Model):
     fecha_entrega = models.DateField(db_column='fecha_Entrega')  # Field name made lowercase.
     estado = models.TextField()
     cod_inventario = models.ForeignKey('Inventario', models.DO_NOTHING, db_column='cod_inventario')
-    cantidad = models.IntegerField()
+    cantidad = models.DecimalField(max_digits=10, decimal_places=2)
     id_orden = models.ForeignKey(OrdenProduccion, models.DO_NOTHING, db_column='id_orden')
     prioridad = models.TextField()
     fecha_creacion = models.DateField()
